Fix for CVE-2021-44228
You won't exploit me!

Obviously, this is a joke commit. This code does not accept external
output, that would be logged during its running. And even if it would,
there is nothing to exploit here. However, it's still best practice to
keep your dependencies up-to-date, in the off chance that somebody will
copy that part of the POM from exactly this project...